# Perf: precompute width and height calculations