With this change, factors for width and height calculations are precomputed in the constructor. This prevents their recalculation every time getPoint is called.

In the `ImagePainter`, where `getPoint()` is executed tens of thousands of times, this can make a significant difference.

Benchmark Results for `ImageShapeBench`:

```
+-----------+---------------------+--------+------------+--------+------------+------+-----------+----------+
| suite     | date                | php    | vcs branch | xdebug | iterations | revs | mode      | net_time |
+-----------+---------------------+--------+------------+--------+------------+------+-----------+----------+
| <current> | 2023-11-11 18:56:52 | 8.2.12 | painter    | false  | 4          | 100  | 109.289ms | 10.898s  |
| main      | 2023-11-11 18:50:03 | 8.2.12 | main       | false  | 4          | 100  | 120.271ms | 12.091s  |
+-----------+---------------------+--------+------------+--------+------------+------+-----------+----------+
```

```
<current>
+-----------------+--------------------+----------+-----------+-----------+-----------+--------+-----------+
| benchmark       | subject            | memory   | min       | max       | mode      | rstdev | stdev     |
+-----------------+--------------------+----------+-----------+-----------+-----------+--------+-----------+
| ImageShapeBench | benchImageShape () | 12.327mb | 108.381ms | 109.518ms | 109.289ms | ±0.44% | 480.182μs |
+-----------------+--------------------+----------+-----------+-----------+-----------+--------+-----------+

main
+-----------------+--------------------+----------+-----------+-----------+-----------+--------+---------+
| benchmark       | subject            | memory   | min       | max       | mode      | rstdev | stdev   |
+-----------------+--------------------+----------+-----------+-----------+-----------+--------+---------+
| ImageShapeBench | benchImageShape () | 12.327mb | 120.072ms | 122.903ms | 120.271ms | ±0.96% | 1.162ms |
+-----------------+--------------------+----------+-----------+-----------+-----------+--------+---------+
```

**Note:**  I have removed this check:

```php
if ($width === 0.0 || $height === 0.0) {
    return null;
}
```

I'm wondering if it has any real use case. Does it? If so, I can bring it back.