@dantleech 2023-11-11
The 0 checks are to catch division by zero