@KennedyTedesco 2023-11-11
@dantleech Yes, but I thought this would never happen in a real case. Bringing back this check made things too verbose (I've pushed the changes here), I think. So, I'm closing this PR for now.