@dantleech 2023-11-11
Probably this bottleneck will become negligible in most cases with #85